chore: relocate reporting file
