+++ src/render/index.js
@@ -1,11 +1,11 @@
+import echo from '../utils/console.js';
+import db, { AUCTION_STORE } from '../utils/db.js';
+import validate from '../utils/validation.js';
 import {
 	getBuyerReport,
 	getSellerReport,
 	hasRendered,
-} from '../reporting/utils.js';
-import echo from '../utils/console.js';
-import db, { AUCTION_STORE } from '../utils/db.js';
-import validate from '../utils/validation.js';
+} from './reporting.js';
 import {
 	getTarget,
 	renderFrame,
